Add files via upload
Finally, the wall functions work. Now I will populate the walls arrays with respective coordinates and draw the Map. This might take a week as the map consists of 70+ subsections. It will be lengthy but definitely worth it. Once I'm done drawing my maps I 'll start working on weapons and shooting